Simplifying and bug-finding.
Changed the agent_state value checks.
Removing the wait_for_initialization delay.

Bug:
The stabilization delay is problematic - it injects up to 60s of sleep into many code-paths. But it can only be
called once the link has been established so it makes the reconnection behavior problematic. There is no
synchronisation over the connected flag and it could change during query processing. The return codes from
query processing are not checked and so messages are dropped if the link changed state.

Bug:
All queries are ignored during popcorning, but commands from the cloud are in the same queue for processing so
the agent would ignore requests during this period.

Bug:
Any query resets the popcorning timer (including encapsulated queries) so popcorning can go on indefinitely.

Bug:
The implementation of the time-out / check is wrong and it will always sleep 10s longer than necessary.