Fix crash.
Worrying that that compiled. Wrong version has been tested (hence the fix to the docker-compose) and something
broke during the cleanup. The crash was caused by calling mosquitto before initialization, this has been
resolved. There is still a problem with the callback ordering that I will fix in the next commit.