Fixing connection.
After the testing problem was fixed and we recovered from the crash the connection stopped working. Investigation
shows that the assumptions about the event loop do not line up in the code. The event loops in the external
library were returning immediately (they only sample their internal queues). The code in the main loop was assuming
that the unused timeout made them block for a second. This combination meant that we were busy-polling the event
loop in the web-socket library and it would not connect. It is still not clear why calling the event loop too
often causes it to fail.

This version inserts an arbitrary delay of 100ms into the main-loop between calls to the event loops. While this has
fixed the connection problem it is ugly and will be problematic.